moved sent resonator notification channels into sub-folder
