+++ src/scheduler/channels/index.js
@@ -1,0 +1,4 @@
+import sendResonatorMail from "./mail";
+import { sendResonatorNotification } from "./push";
+
+export { sendResonatorMail, sendResonatorNotification };
